here is the first file for a text game
here is the first file for the game - it has a very rudimentary Room object, which could be expanded in many possible directions.

I would anticipate a database of rooms to describe a game scenario.  a separate function might be written to function as a parser.  I would like to link an Elisa type natural language pseudo-AI to react to inputs not covered by the base game.